requirements.txt
Django==3.2.6
django-crispy-forms==1.12.0
django-crum==0.7.9
django-filter==2.4.0
django-simple-history==3.0.0
psycopg2-binary==2.9.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
twilio==6.63.1
mysqlclient
